simple network is designed to provide basic connectivity between two computers on separate local networks.
Each computer is connected to a switch, which acts as a central hub for the local network. When a computer wants to send data to another computer on the network, it sends the data to the switch, which then forwards it to the destination computer.

The two switches are connected together, which allows data to be passed between the two local networks. When a computer on one switch wants to communicate with a computer on the other switch, the data is sent to its switch, which then forwards it to the other switch, and ultimately to the destination computer.